refactor(sentry): replace Switcher class with Feature class
- Introduce new Feature class with the same functionality as Switcher
- Update all aspect classes and listeners to use Feature instead of Switcher
- Convert Switcher to extend Feature for backward compatibility with deprecation warnings
- Add appropriate @deprecated annotations for all Switcher methods
- Update test to use Feature class instead of Switcher
- This change improves code clarity while maintaining backward compatibility

Breaking change: Switcher class is deprecated and will be removed in v3.2
